refactor : 패키지 구조 변경
redis는 infra 영역으로 옮김.
