+++ src/main/java/com/meoguri/linkocean/internal/user/application/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.internal.user.domain;
+package com.meoguri.linkocean.internal.user.application;
 
 public interface RefreshTokenService {
