--- src/main/java/com/meoguri/linkocean/internal/user/domain/RefreshTokenRepository.java
@@ -1,8 +1,0 @@
-package com.meoguri.linkocean.internal.user.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.meoguri.linkocean.internal.user.domain.model.RefreshToken;
-
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
-}
