+++ src/main/java/com/meoguri/linkocean/internal/user/infrastructure/redis/RedisRefreshTokenRepository.java
@@ -1,0 +1,6 @@
+package com.meoguri.linkocean.internal.user.infrastructure.redis;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface RedisRefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+}
